- Default exports for baseCoin
